Identification now handle: * Two objects temporarily becoming one (occlusion handling, using prediction), they continue within this new "parent object" as child objects. Children are forced to be within the virtual parent object's boundary. * Objects that becomes hidden temporarily, they continue in their predicted path as lost objects * Objects that move outside or are about to (predicted velocity) are removed. * Most probable objects selected when no overlap is found from previous objects. The error function uses distance (based on predicted position) as well as are difference)
Significant improvements, but a sufficiently sophisticated background model as well as a tuned foreground segmentation is needed for optimum results. Right now a lot of wrongly found objects are deemed lost and stays in the picture when "hard parts" are insufficiently handled by the background model and by part of the foreground segmentation.